Adds support for useful metadata.
The model protos include space for an optional metadata string.
We exploit this in the case restorer to store a table mapping tokens
that take a mixed case (ever) onto their mixed-cased forms, as JSON.
This is an acceptable strategy because, at least in WSJ, there are no
tokens which exhibit multiple mixed-case patterns.

We also expand the support for readable metadata to the other two
working applications, though they just log the metadata string if it's
present.